% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darknessareas_df.R
\docType{data}
\name{darknessareas_df}
\alias{darknessareas_df}
\title{Zeeland Darkness Areas dataframe format}
\format{
\subsection{\code{darknessareas_df}}{

A data frame with 7 rows and 2 columns:
\describe{
\item{WKT}{SF polygon coordinates}
\item{name}{Name of specific darkness area}
}
}
}
\source{
\url{https://zoek.officielebekendmakingen.nl/prb-2019-180.html}
}
\usage{
darknessareas_df
}
\description{
This dataframe contains the coordinates and names of the designated
darkness areas in Zeeland
}
\keyword{datasets}
