% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darknessareas_sf.R
\docType{data}
\name{darknessareas_sf}
\alias{darknessareas_sf}
\title{Zeeland Darkness Areas dataframe format}
\format{
\subsection{\code{darknessareas_sf}}{

An sf object with polygon shapes for each darkness area
}
}
\source{
\url{https://zoek.officielebekendmakingen.nl/prb-2019-180.html}
}
\usage{
darknessareas_sf
}
\description{
This dataframe contains the coordinates and names of the designated
darkness areas in Zeeland
}
\keyword{datasets}
