% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_deflator.R
\docType{data}
\name{gdp_deflator}
\alias{gdp_deflator}
\title{Netherlands GDP Deflator}
\format{
\subsection{\code{gdp_deflator}}{

A data frame with 7 rows and 2 columns:
\describe{
\item{.country_name}{Country Name}
\item{.country_code}{Country Code}
\item{.2015, .2016, .2017, .2018, .2019, .2020, .2021, .2022, .2023}{Years}
}
}
}
\source{
\url{https://data.worldbank.org/indicator/NY.GDP.DEFL.ZS?locations=NL}
}
\usage{
gdp_deflator
}
\description{
This dataframe contains the GDP Deflator for the Netherlands from the
years 2015 to 2023
}
\keyword{datasets}
