% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_regional_nominal.R
\docType{data}
\name{gdp_regional_nominal}
\alias{gdp_regional_nominal}
\title{Netherlands Nominal GDP by Region}
\format{
\subsection{\code{gdp_regional_nominal}}{

A data frame with 7 rows and 2 columns:
\describe{
\item{.region}{Province or Region in the Netherlands}
\item{.2015, .2016, .2017, .2018, .2019, .2020, .2021, .2022}{Years}
}
}
}
\source{
\url{https://opendata.cbs.nl/#/CBS/en/dataset/84432ENG/table}
}
\usage{
gdp_regional_nominal
}
\description{
This dataframe contains the Nominal GDP for the Netherlands by region from the
years 2015 to 2022
}
\keyword{datasets}
