% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_regional_real.R
\docType{data}
\name{gdp_regional_real}
\alias{gdp_regional_real}
\title{Netherlands real GDP by Region}
\format{
\subsection{\code{gdp_regional_real}}{

A data frame with 7 rows and 2 columns:
\describe{
\item{.region}{Province or Region in the Netherlands}
\item{.2015, .2016, .2017, .2018, .2019, .2020, .2021, .2022}{Years}
}
}
}
\source{
\url{https://opendata.cbs.nl/#/CBS/en/dataset/84432ENG/table} \url{https://data.worldbank.org/indicator/NY.GDP.DEFL.ZS?locations=NL}
}
\usage{
gdp_regional_real
}
\description{
This dataframe contains the real GDP for the Netherlands by region from the
years 2015 to 2022
}
\keyword{datasets}
