% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NTLdf.R
\name{get_NTLdf}
\alias{get_NTLdf}
\title{Create NTL dataframe}
\usage{
get_NTLdf(country, bearer, product_id = "VNP46A4", ymd, tidy = TRUE)
}
\arguments{
\item{country}{country of interest}

\item{bearer}{Your personal bearer token from NASA Earthdata}

\item{product_id}{VNP46A1, VNP46A2, VNP46A3, VNP46A4}

\item{ymd}{Year or Y-M-D}

\item{tidy}{when TRUE, outputs a tidied dataframe}

\item{low_rm}{Remove lowest values}
}
\value{
A dataframe object containing NTL values and coordinates
}
\description{
Create NTL dataframe
}
\examples{
get_NTLraster("NLD", "VNP46A4", 2015, "examplebearertoken", "FALSE")
}
