% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NTLmap.R
\name{get_NTLmap}
\alias{get_NTLmap}
\title{Create NTL map}
\usage{
get_NTLmap(raster, maptheme = "Nighttime", ymd)
}
\arguments{
\item{raster}{get_NTLraster raster object}

\item{maptheme}{Nighttime, Neonheatmap or Pastelheatmap}

\item{ymd}{ymd or date of map}
}
\value{
Map plot
}
\description{
Create NTL map
}
\examples{

get_NTLmap(NTLraster, "Nighttime", 2015)
}
