% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NTLraster.R
\name{get_NTLraster}
\alias{get_NTLraster}
\title{Create NTL Raster}
\usage{
get_NTLraster(country, product_id, ymd, low.rm = FALSE, mask = FALSE)
}
\arguments{
\item{country}{country of interest}

\item{product_id}{VNP46A1, VNP46A2, VNP46A3, VNP46A4}

\item{ymd}{Year or Y-M-D}

\item{bearer}{Your personal bearer token from NASA Earthdata}

\item{remlow}{Remove lowest values}
}
\value{
A raster object
}
\description{
Create NTL Raster
}
\examples{
get_NTLraster("NLD", "VNP46A4", 2015, "examplebearertoken", "FALSE")
}
