% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NTLrastermasked.R
\name{get_NTLrastermasked}
\alias{get_NTLrastermasked}
\title{Create NTL Raster masked}
\usage{
get_NTLrastermasked(
  country = "NLD",
  bearer,
  product_id,
  ymd,
  output_raster = TRUE
)
}
\arguments{
\item{country}{country of interest}

\item{bearer}{Your personal bearer token from NASA Earthdata}

\item{product_id}{VNP46A1, VNP46A2, VNP46A3, VNP46A4}

\item{ymd}{Year or Y-M-D}

\item{output_raster}{When FALSE, outputs a dataframe object}
}
\value{
A raster or dataframe object for plotting masked by country
}
\description{
Create NTL Raster masked
}
\examples{
get_NTLrastermasked("NLD", "examplebearertoken", "VNP46A4", 2015, "FALSE")
}
