% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NTLspatdf.R
\name{get_NTLspatdf}
\alias{get_NTLspatdf}
\title{Create NTL spatial points dataframe with necessary transformations (for plotting with ggplot2)}
\usage{
get_NTLspatdf(country = "NLD", bearer, product_id = "VNP46A4", ymd)
}
\arguments{
\item{country}{country of interest}

\item{bearer}{Your personal bearer token from NASA Earthdata}

\item{product_id}{VNP46A1 (Raw Daily), VNP46A2 (Adjusted Daily), VNP46A3(Monthly), VNP46A4(Yearly)}

\item{ymd}{Year or Y-M-D}
}
\value{
A spatial dataframe object for plotting
}
\description{
Create NTL spatial points dataframe with necessary transformations (for plotting with ggplot2)
}
\examples{
get_NTLspatdf("NLD", "examplebearertoken", "VNP46A4", 2015)

}
