% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_NTLgif.R
\name{save_NTLgif}
\alias{save_NTLgif}
\title{Create NTL Raster masked}
\usage{
save_NTLgif(pathimages, pathsave)
}
\arguments{
\item{pathimages}{folder path of images to be saved as GIF}

\item{pathsave}{folder path to save gif}
}
\value{
A raster or dataframe object for plotting masked by country
}
\description{
Create NTL Raster masked
}
