% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_NTLimages.R
\name{save_NTLimages}
\alias{save_NTLimages}
\title{Create NTL Raster masked}
\usage{
save_NTLimages(
  country = "NLD",
  bearer,
  product_id = "VNP46A4",
  years,
  maptheme = "PastelHeatMap",
  folderpath
)
}
\arguments{
\item{country}{country of interest}

\item{bearer}{Your personal bearer token from NASA Earthdata}

\item{product_id}{VNP46A1, VNP46A2, VNP46A3, VNP46A4}

\item{years}{Vector of years}

\item{maptheme}{"Nighttime" "NeonHeatMap" or "PastelHeatMap"}
}
\value{
A raster or dataframe object for plotting masked by country
}
\description{
Create NTL Raster masked
}
\examples{
get_NTLrastermasked("NLD", "examplebearertoken", "VNP46A4", 2015, "FALSE")
}
