% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_neonheatmap.R
\name{theme_neonheatmap}
\alias{theme_neonheatmap}
\title{Add neonheatmap theme to plot}
\usage{
theme_neonheatmap()
}
\value{
A plot object with theme applied
}
\description{
Add neonheatmap theme to plot
}
\examples{
plot + theme_neonheatmap

}
