% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nighttime.R
\name{theme_nighttime}
\alias{theme_nighttime}
\title{Add nighttime theme to plot}
\usage{
theme_nighttime(plot)
}
\arguments{
\item{plot}{a ggplot object}
}
\value{
A plot object with theme applied
}
\description{
Add nighttime theme to plot
}
\examples{
plot |> theme_nighttime

}
