% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pastelheatmap.R
\name{theme_pastelheatmap}
\alias{theme_pastelheatmap}
\title{Add pastelheatmap theme to plot}
\usage{
theme_pastelheatmap()
}
\value{
A plot object with theme applied
}
\description{
Add pastelheatmap theme to plot
}
\examples{
plot + theme_pastelheatmap

}
