% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NTLmap.R
\name{get_NTLmap}
\alias{get_NTLmap}
\title{Create NTL map to use in combination with themes}
\usage{
get_NTLmap(raster, ymd)
}
\arguments{
\item{ymd}{ymd or date of map}

\item{spatialdf}{spatial dataframe object from get_spatdf}
}
\value{
Map plot
}
\description{
Create NTL map to use in combination with themes
}
\examples{

get_NTLmap(NTLraster, 2015)
}
