% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NTLraster.R
\name{get_NTLraster}
\alias{get_NTLraster}
\title{Create NTL Raster}
\usage{
get_NTLraster(
  country = "NLD",
  bearer,
  product_id,
  ymd,
  low_rm = FALSE,
  mask = FALSE,
  output_type
)
}
\arguments{
\item{country}{country of interest}

\item{bearer}{Your personal bearer token from NASA Earthdata}

\item{product_id}{VNP46A1, VNP46A2, VNP46A3, VNP46A4}

\item{ymd}{Year or Y-M-D}

\item{low_rm}{Remove lowest values}

\item{raster_to_points}{When TRUE outputs datarfame}
}
\value{
A raster or dataframe object for plotting
}
\description{
Create NTL Raster
}
\examples{
get_NTLraster("NLD", "VNP46A4", 2015, "examplebearertoken", "FALSE")
}
