% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NTLraster.R
\name{get_NTLraster}
\alias{get_NTLraster}
\title{Create NTL Raster with necessary transformations}
\usage{
get_NTLraster(country = "NLD", bearer, product_id = "VNP46A4", ymd)
}
\arguments{
\item{country}{country of interest}

\item{bearer}{Your personal bearer token from NASA Earthdata}

\item{product_id}{VNP46A1 (Raw Daily), VNP46A2 (Adjusted Daily), VNP46A3(Monthly), VNP46A4(Yearly)}

\item{ymd}{Year or Y-M-D}
}
\value{
A raster object for plotting
}
\description{
Create NTL Raster with necessary transformations
}
\examples{
get_NTLraster("NLD", "examplebearertoken", "VNP46A4", 2015, "FALSE")

}
