% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stackvalues.R
\name{get_stackvalues}
\alias{get_stackvalues}
\title{Create NTL Raster}
\usage{
get_stackvalues(rasterstack)
}
\arguments{
\item{rasterstack}{a rasterstack or rasterlayer object}
}
\value{
A dataframe with each layer and its values
}
\description{
Create NTL Raster
}
\examples{
get_NTLraster(rasterstack)

}
